SalaisetReseptit/src/main/java/recipes/domain/User.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package recipes.domain;

/**
 *
 * @author aebjork
 */
public class User {

    private String firstname;
    private String lastname;
    private String email;
    private String username;
    private String password;

    public User(String firstname, String lastname, String email, String username, String password) {
        this.firstname = firstname;
        this.lastname = lastname;
        this.email = email;
        this.username = username;
        this.password = password;
    }

    public String getFirstname() {
        return this.firstname;
    }

    public String getLastname() {
        return this.lastname;
    }

    public String getEmail() {
        return this.email;
    }

    public String getUsername() {
        return this.username;
    }

    public String getPassword() {
        return this.password;
    }

    @Override
    public String toString() {
        return "\n" + firstname + " " + lastname
                + "\n" + email + "\n" + username
                + "\n" + password;
    }

}
